requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
pillow
scikit-learn
gitpython
gitdb
defusedxml
ipython
future
zlib
jupyter
pip
panel
attrs
distro
wheel
pysocks
protobuf
openssl
pyopenssl
cryptography
docutils
sphinx
filelock
pyzmq
colorama
pexpect
pytest
tornado
comm
traitlets
parso
jedi
pytz
rich
pygments
markdown-it-py
toml
zipp
typing_extensions
gettext
smmap
mdurl
charset-normalizer
setuptools
overrides
cffi
numba
mypy
scipy
threadpoolctl
blas
mpmath
sympy
platformdirs
altair
jsonschema
pyarrow
jinja2
packaging
referencing
markupsafe
joblib
lz4
distributed
dask
psutil
cloudpickle
xlrd
openpyxl
fsspec
sqlalchemy
lxml
numexpr
python-dateutil
tenacity
xarray
ipywidgets
scikit-image
pydeck
certifi
debugpy
greenlet
decorator
regex
stack_data
matplotlib-inline
pickleshare
jupyter_client
prompt_toolkit
prompt-toolkit
wcwidth
ptyprocess
tzdata
cycler
statsmodels
twisted
pycurl
zstandard
urllib3
six
dask-expr
idna
chardet
astroid
asttokens
contourpy
bokeh
executing
fonttools
freetype
cairo
unicodedata2
ipykernel
jupyter_core
pure_eval
pyparsing
click
cachetools
blinker
watchdog
graphviz
jsonpointer
rfc3339-validator
jsonschema-specifications
kiwisolver
linkify-it-py